// src/app/models/Reports-page/PenaltyReportModel.tsx
export interface IPenaltyReportModel {
  penaltyID: number
  projectID: number
  empVendorID: number
  projectName: string
  empVendorName: string
  customerName: string
  penaltyByName: string
  mobileNumber: string
  remarks: string
  penaltyTypeName: string
  penaltyForName: string
  status: true
  penaltyCreateDate: string
  responseDate: string
  amount: boolean
  departmentName: string
  approveForName: string
}